refactor: decouple user info repository
