+++ src/libs/modules/database/repositories/userInfo/userInfo.interface.ts
@@ -1,0 +1,3 @@
+export interface UserInfoRepository {}
+
+export const UserInfoRepository = Symbol('UserInfoRepository');
